requirements.txt
langchain
langchain-openai
openai
huggingface_hub
langchain-huggingface
langchain-community
python-dotenv
streamlit